Netcat funtion
This is a way for one system to serve their /bin/bash (or any other process/path) on port 3333 so that the "other" player can make changes on the serving player's behalf. -v is for verbose output. -l means localhost address. -p source port. -e executes a given command. This is using ncat which is a nmap implementation of netcat. Is this a bad dependency?